server/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
spandrel
pillow
numpy
